requirements.txt
fastapi>=0.110
uvicorn>=0.29
crewai[bedrock]>=0.28.8
litellm>=1.40
python-hcl2>=4.3
pyyaml>=6.0
feedparser>=6.0
anthropic>=0.49
boto3>=1.28.57
pydantic>=2.0
pydantic-settings>=2.0
requests>=2.31
python-dotenv>=1.0
python-multipart>=0.0.9
httpx>=0.27.0
